Support CYPRESS_RUN_BINARY
